KL-149/feat: add CategoryService interface
